requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
joblib
matplotlib
seaborn
beautifulsoup4
requests
jinja2
pyyaml
shap
tabpfn
pandoc 
